module6-task1 исправление замечаний
