+++ data/data_for_auth.php
@@ -1,0 +1,4 @@
+<?php
+$is_auth = rand(0, 1);
+$user_name = 'Евгений';
+$title = "readme: популярное";
